requirements.txt
torch==1.1.0
numpy==1.17.2
scikit-learn==0.20.1
pandas==1.1.0
tqdm==4.31.1
torchvision==0.3.0
arrow==0.17.0
logzero==1.5.0
PyYAML==5.1.1
tensorboard==1.14.0
future==0.16.0
